Use git checkout so version checks work again, fix up logic in start script

+++ run-tautulli.sh
@@ -10,11 +10,11 @@ if [ "$( id -g tautulli )" -ne "${LGID}" ]; then
 fi
 
 # Move the old config if it exists
-if [ ! -f "/config/plexpy.ini" ]; then
+if [ -f "/config/plexpy.ini" ] && [ ! -f "/config/tautulli.ini" ]; then
     mv /config/plexpy.ini /config/tautulli.ini
 fi
 
 # Set permissions
 chown -R tautulli:tautulli /config/ /opt/tautulli
 
-exec runuser -l tautulli -c '/opt/tautulli/PlexPy.py --nolaunch --quiet --config /config/tautulli.ini --datadir /config/data'
+exec runuser -l tautulli -c '/opt/tautulli/app/PlexPy.py --nolaunch --quiet --config /config/tautulli.ini --datadir /config/data'
